Maximun Difference In An Array
The maximum difference for a pair of elements in some array a is defined as the largest difference between any a[i] and a[j] where i < j and a[i] < a[j]. The declaration for a function named maxDifference (which takes array a as a parameter) is provided for you in the editor. Complete the function so that it calculates and returns the maximum difference for a; if no such number exists (e.g.: if a is in descending order and all a[j] < a[i]), return -1. Input Format The task of reading input (defined below) from stdin, assembling it into array a, and and calling maxDifference(a) is already handled for you by the locked code in the editor. The first line contains N (the number of elements in array a). The N subsequent lines each contain a single element of a; the ith line of input (where 0 < i < N-1) contains element a[i].